src/third_party/blink/web_tests/external/wpt/css/css-scrollbars/scrollbar-width-keywords-expected.txt

This is a testharness.js-based test.
FAIL scrollbar-width: none should have no scrollbar assert_equals: expected 100 but got 85
PASS scrollbar-width: thin should have scrollbar no wider than auto
Harness: the test ran to completion.
